# docs/swagger.yaml
definitions:
  models.AddRoleRequest:
    properties:
      name:
        type: string
    required:
    - name
    type: object
  models.CreateNewAccessTokenRequest:
    properties:
      refresh_token:
        type: string
    type: object
  models.CreateTransactionRequest:
    properties:
      amount:
        type: number
      payment_method:
        type: string
    type: object
  models.ResetPasswordRequest:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  models.SendNewOTPRequest:
    properties:
      email:
        type: string
    type: object
  models.SignInRequest:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  models.SignUpRequest:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        type: string
    type: object
  models.UpdatePasswordRequest:
    properties:
      password:
        type: string
    type: object
  models.UpdateProfileRequest:
    properties:
      email:
        type: string
      image:
        type: string
      name:
        type: string
      phone_number:
        type: string
    type: object
  models.UpdateTransactionStatusRequest:
    properties:
      status:
        type: string
    type: object
  models.VerifyAccountRequest:
    properties:
      email:
        type: string
      otp_code:
        type: string
    type: object
info:
  contact: {}
paths:
  /api/auth/create-new-access-token:
    post:
      consumes:
      - application/json
      description: Create new access token by refresh token
      parameters:
      - description: User data
        in: body
        name: userData
        required: true
        schema:
          $ref: '#/definitions/models.CreateNewAccessTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "401":
          description: Unauthorized
        "500":
          description: Internal Server Error
      summary: Create new access token
      tags:
      - Authentication
  /api/auth/signin:
    post:
      consumes:
      - application/json
      description: Login with email and password
      parameters:
      - description: User data
        in: body
        name: userData
        required: true
        schema:
          $ref: '#/definitions/models.SignInRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "401":
          description: Unauthorized
        "500":
          description: Internal Server Error
      summary: Login as an authenticated user
      tags:
      - Authentication
  /api/role:
    get:
      consumes:
      - application/json
      description: Get all user roles
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Get roles
      tags:
      - Roles
    post:
      consumes:
      - application/json
      description: Create new role
      parameters:
      - description: User data
        in: body
        name: userData
        required: true
        schema:
          $ref: '#/definitions/models.AddRoleRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
        "400":
          description: Bad Request
        "409":
          description: Conflict
        "500":
          description: Internal Server Error
      summary: Add Role
      tags:
      - Roles
  /api/transactions:
    get:
      consumes:
      - application/json
      description: Retrieve transactions with optional filters (user_id, status, pagination)
      parameters:
      - description: User ID filter
        in: query
        name: user_id
        type: string
      - description: Transaction status filter (pending, success, failed)
        in: query
        name: status
        type: string
      - description: 'Page number for pagination (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Limit number of transactions per page (default: 10)'
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successful response
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get Transactions
      tags:
      - Transactions
    post:
      consumes:
      - application/json
      description: Create transaction
      parameters:
      - description: Authorization token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Transaction data
        in: body
        name: userData
        required: true
        schema:
          $ref: '#/definitions/models.CreateTransactionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "500":
          description: Internal Server Error
      summary: Create Transaction
      tags:
      - Transactions
  /api/transactions/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a transaction by role ID
      parameters:
      - description: Transaction ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Delete a transaction
      tags:
      - Transactions
    get:
      consumes:
      - application/json
      description: Fetch the transaction details based on the ID provided
      parameters:
      - description: Transaction ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Get transaction by ID
      tags:
      - Transactions
    put:
      consumes:
      - application/json
      description: Update transaction status
      parameters:
      - description: Transaction ID
        in: path
        name: id
        required: true
        type: string
      - description: Transaction data
        in: body
        name: userData
        required: true
        schema:
          $ref: '#/definitions/models.UpdateTransactionStatusRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: Update Transaction
      tags:
      - Transactions
  /api/transactions/dashboard/summary:
    get:
      consumes:
      - application/json
      description: Retrieve transactions summary
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: Get Transactions summary
      tags:
      - Transactions
  /api/user-service/role/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a role by role ID
      parameters:
      - description: Role ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Delete a role
      tags:
      - Roles
    get:
      consumes:
      - application/json
      description: Get role by role ID
      parameters:
      - description: Role ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Get a role
      tags:
      - Roles
  /api/user/:
    get:
      consumes:
      - application/json
      description: Fetch all users
      parameters:
      - description: Authorization token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Get all users
      tags:
      - Users
  /api/user/{id}:
    get:
      consumes:
      - application/json
      description: Fetch the user details based on the ID provided
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: Authorization token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "401":
          description: Unauthorized
        "500":
          description: Internal Server Error
      summary: Get user by ID
      tags:
      - Users
  /api/user/forgot-password-otp:
    put:
      consumes:
      - application/json
      description: Resend OTP code for password reset
      parameters:
      - description: User data
        in: body
        name: userData
        required: true
        schema:
          $ref: '#/definitions/models.SendNewOTPRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Send OTP code
      tags:
      - Users
  /api/user/get-user-by-email:
    get:
      consumes:
      - application/json
      description: Get user data by email
      parameters:
      - description: User email
        in: query
        name: email
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "401":
          description: Unauthorized
        "500":
          description: Internal Server Error
      summary: Get user by email
      tags:
      - Users
  /api/user/profile:
    get:
      consumes:
      - application/json
      description: Fetch the user profile based on the decoded ID from access token
      parameters:
      - description: Authorization token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "401":
          description: Unauthorized
        "500":
          description: Internal Server Error
      summary: Get user profile
      tags:
      - Users
  /api/user/reset-password:
    put:
      consumes:
      - application/json
      description: Reset Password
      parameters:
      - description: User data
        in: body
        name: userData
        required: true
        schema:
          $ref: '#/definitions/models.ResetPasswordRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Reset Password
      tags:
      - Users
  /api/user/send-otp:
    put:
      consumes:
      - application/json
      description: Reusable OTP code sender
      parameters:
      - description: User data
        in: body
        name: userData
        required: true
        schema:
          $ref: '#/definitions/models.SendNewOTPRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Send OTP code
      tags:
      - Users
  /api/user/signup:
    post:
      consumes:
      - application/json
      description: Register a new user with name, email, and password
      parameters:
      - description: User data
        in: body
        name: userData
        required: true
        schema:
          $ref: '#/definitions/models.SignUpRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
        "400":
          description: Bad Request
        "409":
          description: Conflict
        "500":
          description: Internal Server Error
      summary: Register a new user
      tags:
      - Users
  /api/user/update:
    put:
      consumes:
      - application/json
      description: Update user profile
      parameters:
      - description: Authorization token
        in: header
        name: Authorization
        required: true
        type: string
      - description: User data
        in: body
        name: userData
        required: true
        schema:
          $ref: '#/definitions/models.UpdateProfileRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Update user profile
      tags:
      - Users
  /api/user/update-password:
    put:
      consumes:
      - application/json
      description: Update user password
      parameters:
      - description: Authorization token
        in: header
        name: Authorization
        required: true
        type: string
      - description: User data
        in: body
        name: userData
        required: true
        schema:
          $ref: '#/definitions/models.UpdatePasswordRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Update user password
      tags:
      - Users
  /api/user/verify-account:
    put:
      consumes:
      - application/json
      description: Account verification after registering a new account
      parameters:
      - description: User data
        in: body
        name: userData
        required: true
        schema:
          $ref: '#/definitions/models.VerifyAccountRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "500":
          description: Internal Server Error
      summary: Account verification
      tags:
      - Users
swagger: "2.0"